% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulledRates-package.R
\docType{package}
\name{pulledRates-package}
\alias{pulledRates}
\alias{pulledRates-package}
\title{pulledRates: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several 
    lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ajhelmstetter/pulledRates}
  \item Report bugs at \url{https://github.com/ajhelmstetter/pulledRates/issues}
}

}
\author{
\strong{Maintainer}: Andrew Helmstetter \email{andrew.j.helmstetter@gmail.com} (\href{https://orcid.org/0000-0003-3761-4981}{ORCID})

}
\keyword{internal}
